major backend refactor.  moviong code out of main and into their own files for ease of collaboration.  made a test data file to fill with various payloads to use in testing.  tried to remove as many supabase calls as possible out of the individual files and into the middleman.
